+++ src/client/mash/IndexBufferManager.hpp
@@ -36,7 +36,7 @@ public:
             - suitable memory type is not found,
             - memory mapping fails.
     */
-    IndexBufferManager(const std::vector<uint16_t> indices, VkCommandPool commandPool);
+    IndexBufferManager(const std::vector<uint32_t> indices, VkCommandPool commandPool);
 
     /*
     @brief Destructor. Cleans up the index buffer and its allocated memory.
